(PUP-10439) Replace Puppet::Network::HTTP::Connection with adapter
The Connection and ConnectionAdapter existed because there were two
different HTTP implementations. In Puppet 7, the former implementation
is being removed, so it is no longer necessary to keep the Connection
implementation.

We could delete the Connection class, but some callers assume
HttpPool.http_instance returns an instance of the Connection class. We
could alias the Connection constant to be the same as ConnectionAdapter,
but I think the simplest thing is to collapse the two classes, keeping
the Connection class.
