+++ lib/puppet/network/http_pool.rb
@@ -1,5 +1,4 @@
 require 'puppet/network/http/connection'
-require 'puppet/network/http/connection_adapter'
 require 'puppet/util/platform'
 
 module Puppet::Network; end
@@ -13,7 +12,7 @@ module Puppet::Network; end
 #
 module Puppet::Network::HttpPool
 
-  @http_client_class = Puppet::Network::HTTP::ConnectionAdapter
+  @http_client_class = Puppet::Network::HTTP::Connection
 
   def self.http_client_class
     @http_client_class
