+++ lib/puppet/runtime.rb
@@ -11,8 +11,7 @@ class Puppet::Runtime
     @runtime_services = {
       http: proc do
         klass = Puppet::Network::HttpPool.http_client_class
-        if klass == Puppet::Network::HTTP::Connection ||
-           klass == Puppet::Network::HTTP::ConnectionAdapter
+        if klass == Puppet::Network::HTTP::Connection
           Puppet::HTTP::Client.new
         else
           Puppet::HTTP::ExternalClient.new(klass)
